fixing npe when not entering the weighting tab, also fixed the weighting being incorrect

+++ src/main/java/com/ouchadam/fyp/presentation/RuleContainer.java
@@ -6,9 +6,9 @@ public class RuleContainer<T> {
 
     public final FitnessRule<T> fitnessRule;
     public final RuleName ruleName;
-    public final int weight;
+    public final float weight;
 
-    public RuleContainer(FitnessRule<T> fitnessRule, RuleName ruleName, int weight) {
+    public RuleContainer(FitnessRule<T> fitnessRule, RuleName ruleName, float weight) {
         this.fitnessRule = fitnessRule;
         this.ruleName = ruleName;
         this.weight = weight;
